REF: Move references from pospred to position
