+++ ptychonn/position/position_list.py
@@ -4,7 +4,7 @@ import numpy as np
 import matplotlib.pyplot as plt
 import matplotlib
 
-from ptychonn.pospred.io import load_probe_positions_from_file, save_positions_to_csv
+from ptychonn.position.io import load_probe_positions_from_file, save_positions_to_csv
 
 class ProbePositionList:
     def __init__(self, file_path=None, position_list=None, unit='pixel', psize_nm=None, convert_to_pixel=True,
